+++ src/views/SignDidExtrinsic/SignDidExtrinsic.tsx
@@ -1,7 +1,5 @@
-import { Fragment, ReactNode, useCallback, useEffect, useState } from 'react';
+import { Fragment, useCallback } from 'react';
 import { browser } from 'webextension-polyfill-ts';
-import { map } from 'lodash-es';
-import { GenericCall } from '@polkadot/types';
 import { BlockchainApiConnection } from '@kiltprotocol/chain-helpers';
 import { BaseDidKey } from '@kiltprotocol/types';
 
@@ -19,56 +17,7 @@ import {
 import { backgroundSignDidExtrinsicChannel } from '../../channels/SignDidExtrinsicChannels/backgroundSignDidExtrinsicChannel';
 import { SignDidExtrinsicOriginInput } from '../../channels/SignDidExtrinsicChannels/types';
 
-interface Value {
-  value: ReactNode;
-  label: string;
-}
-
-export function useExtrinsicValues(
-  input: SignDidExtrinsicOriginInput,
-): Value[] {
-  const [values, setValues] = useState<Value[]>([]);
-
-  useEffect(() => {
-    (async () => {
-      const t = browser.i18n.getMessage;
-
-      const { origin } = input;
-      const { api } = await BlockchainApiConnection.getConnectionOrConnect();
-
-      const extrinsic = api.createType('Extrinsic', input.extrinsic);
-
-      const { meta } = extrinsic;
-      const { section } = extrinsic.method;
-
-      const forbidden = section === 'did';
-      const errorLine = {
-        label: 'FORBIDDEN',
-        value: 'All did.* calls are forbidden',
-      };
-      const error = !forbidden ? [] : [errorLine];
-
-      const argumentValues = (extrinsic.method as GenericCall).toHuman()
-        .args as Record<string, unknown>;
-      const argumentNames = meta && map(meta.args, 'name');
-      const nameValuePairs = argumentNames.map(
-        (name) =>
-          `${name} = ${JSON.stringify(argumentValues[name.toString()])}`,
-      );
-      const methodSignature = meta ? `(${nameValuePairs.join(', ')})` : '';
-
-      const method = `${section}.${extrinsic.method.method}${methodSignature}`;
-
-      setValues([
-        ...error,
-        { value: origin, label: t('view_SignDidExtrinsic_from') },
-        { value: method, label: t('view_SignDidExtrinsic_method') },
-      ]);
-    })();
-  }, [input]);
-
-  return values;
-}
+import { useExtrinsicValues } from './useExtrinsicValues';
 
 interface Props {
   identity: Identity;
